Use new size hints for vectors
This is much more precise encoding with both lower and upper bound. It
implements idea discussed in #388 and for example avoids problems from #301.

However benchmarks result are at best mixed: benchmarks change range from 0.75
to 17. Investigation of tridiag benchmark (it's not worst but one of simplest)
showed that main loop retained Bundles, allocated closures in inner loop and so
were quite slow.

It seems that generation of tight loops from vector functions is rather fragile
and what worse we have no way to know whether this problem exists for code in
the wild and have no way to measure this.